requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
beautifulsoup4==4.12.3
requests==2.32.2
pydantic==2.7.2
python-dotenv==1.0.1
pandas==2.2.2